Refactored state manager remove APIs
